+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
+- Feature: Add new `Title` React component.
 
 ## [5.3.0] - 2017-01-16